requirements.txt
asgiref==3.2.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
et-xmlfile==1.0.1
filter==0.0.0.20200724
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.3
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.4
packaging==20.9
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
